+++ babel.config.js
@@ -1,3 +1,6 @@
 module.exports = {
-  presets: ['module:@react-native/babel-preset', '@babel/preset-flow']
+  presets: [
+    'module:@react-native/babel-preset',
+    '@babel/preset-typescript'
+  ]
 }