+++ plugin/test-stubs/expo/config.js
@@ -1,0 +1,6 @@
+exports.getNameFromConfig = function(appJSON) {
+  return { appName: appJSON?.name || 'App', webName: appJSON?.name || 'App' };
+};
+exports.getConfig = function(projectRoot) {
+  return { exp: { name: 'App', slug: 'app', web: {}, ios: {}, android: {} } };
+};